Extract Navigation Desktop Menu to a React component
