+++ src/components/Header.tsx
@@ -6,6 +6,7 @@ import {FaMoon, FaSun, FaBars, FaTimes} from 'react-icons/fa'
 import {useState, useEffect, useRef} from 'react'
 import Breadcrumbs from "@/components/Breadcrumbs";
 import {useTheme} from "@/hooks/useTheme";
+import NavigationMenu from "@/components/NavigationMenu";
 
 const navItems = [
     {name: 'Home', path: '/'},
@@ -16,7 +17,6 @@ const navItems = [
 
 export default function Header() {
     const pathname = usePathname()
-    const [activeIndex, setActiveIndex] = useState(0)
     const [mobileMenuOpen, setMobileMenuOpen] = useState(false)
     const menuRef = useRef<HTMLDivElement | null>(null)
     const [theme, setTheme] = useTheme()
@@ -24,13 +24,6 @@ export default function Header() {
     const toggleTheme = () => {
         setTheme(theme === 'dark' ? 'light' : 'dark')
     }
-
-    useEffect(() => {
-        const index = navItems.findIndex(({path}) =>
-            path === '/' ? pathname === '/' : pathname.startsWith(path) // Match `/blog` for `/blog/...`
-        )
-        setActiveIndex(index !== -1 ? index : 0)
-    }, [pathname])
 
     useEffect(() => {
         const handleClickOutside = (event: MouseEvent) => {
@@ -61,32 +54,7 @@ export default function Header() {
                 <Breadcrumbs/>
 
                 {/* Center: Segmented navigation - Hidden on mobile */}
-                <nav className="hidden md:block absolute left-1/2 top-1/2 transform -translate-x-1/2 -translate-y-1/2">
-                    <ul className="flex items-center justify-center gap-1 border border-gray-600 rounded-full px-1 py-1 relative">
-                        <div
-                            className="absolute top-0 left-0 h-full border-2 border-blue-500 rounded-full transition-transform duration-300 pointer-events-none"
-                            style={{
-                                width: `calc(100% / ${navItems.length})`,
-                                transform: `translateX(${activeIndex * 100}%)`,
-                            }}
-                        ></div>
-                        {navItems.map(({name, path}) => {
-                            const isActive = pathname === path
-                            return (
-                                <li key={name} className="relative z-10 flex justify-center items-center">
-                                    <Link
-                                        href={path}
-                                        className={`px-4 py-2 rounded-full text-sm font-medium text-center transition-all ${
-                                            isActive ? 'text-black dark:text-white' : 'text-black dark:text-white dark:hover:bg-gray-800 hover:bg-gray-200'
-                                        }`}
-                                    >
-                                        {name}
-                                    </Link>
-                                </li>
-                            )
-                        })}
-                    </ul>
-                </nav>
+                <NavigationMenu/>
 
                 {/* Right side: Theme toggle + Mobile Menu Toggle */}
                 <div className="flex items-center gap-4">
